Rework c100

+++ tickets/ticket-list.php
@@ -5,17 +5,19 @@
         <div class="text-center mt-5 pt-5">
             <h3>Ticket Liste</h3>
         </div>
+        <input class="form-control" id="myInput" type="text" placeholder="Search..">
 
         <?php
         require_once('../database/config.php');
         include_once('../navbar.php');
         include_once('modal-assign-ticket.php');
+        include_once('modal-delete-ticket.php');
 
         #get DB content
         $sql = "SELECT TicketID,Datum,Abteilung,Name,Problem,Assign FROM ticket_table";
         $result = $link->query($sql);
         ?>
-        <main class="container" style="margin-top: 100px;">
+        <main class="container" style="margin-top: 25px; overflow-y: auto; height: calc(100vh - 210px);">
             <?php
 
             echo "
@@ -27,15 +29,16 @@
               <th scope='col'>Name</th>
               <th scope='col'>Problem</th>
               <th scope='col'>Erfasst am</th>
-              <th scope='col'>Bearbeitet von:</th>
+              <th scope='col'>Bearbeitet von</th>
               <th scope='col'></th>
               <th scope='col'></th>
           </tr>
       </thead>";
 
             while ($row = $result->fetch_assoc()) {
-                $row1 = $row['TicketID'];
+                $ticketID = $row['TicketID'];
                 echo "
+        <tbody id='myTable'>        
         <tr>
             <th scope='row'>$row[TicketID]</th>
             <td>$row[Abteilung]</td>
@@ -43,18 +46,25 @@
             <td>$row[Problem]</td>
             <td>$row[Datum]</td>
             <td>$row[Assign]</td>
-            <td><button class='btn btn-info' data-toggle='modal' data-id='$row1' onclick=\"$('#dataid').text($(this).data('id'));$('#myModal2').modal('show');\">Assign</button></td>
-            <td><a href='delete-ticket.php?id=" . $row['TicketID'] . "'><button class='btn btn-danger' >Delete</button></a></td>
+            <td><button class='btn btn-info' data-toggle='modal' data-id='$ticketID' onclick=\"$('#dataid').text($(this).data('id'));$('#modalAssignTicket').modal('show');\">Assign</button></td>
+            <td><a data-toggle='modal' data-target='#confirm-delete' data-href='delete-ticket.php?id=" . $row['TicketID'] . "'><button class='btn btn-danger' >Delete</button></a></td>
                 </tr>";
             }
             echo "</tbody>
   </table>";
-
-
             ?>
 
         </main>
-
     </div>
-
 </body>
+<!-- Search Ticket Table -->
+<script>
+    $(document).ready(function() {
+        $("#myInput").on("keyup", function() {
+            var value = $(this).val().toLowerCase();
+            $("#myTable tr").filter(function() {
+                $(this).toggle($(this).text().toLowerCase().indexOf(value) > -1)
+            });
+        });
+    });
+</script>
